+++ src/formatters/SimpleCellFormatter.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { FormatterProps } from '../types';
-
-export function SimpleCellFormatter({ row, column }: FormatterProps) {
-  const value = row[column.key];
-  return <span title={String(value)}>{value}</span>;
-}