+++ src/test/GridPropHelpers.ts
@@ -1,8 +1,8 @@
-import { ValueFormatter, SimpleCellFormatter } from '../formatters';
+import { ValueFormatter } from '../formatters';
 import { CalculatedColumn } from '../types';
 
 export interface Row {
-  id?: number;
+  id: number;
   title?: string;
   count?: number;
   description?: string;
@@ -16,7 +16,7 @@ const columns: CalculatedColumn<Row>[] = [{
   left: 0,
   resizable: false,
   sortable: false,
-  formatter: SimpleCellFormatter
+  formatter: ValueFormatter
 }, {
   idx: 1,
   key: 'title',